ref: inline methods without business logic from budget controller
